requirements.txt
pyautogui
pyperclip
anthropic
python-dotenv
google-generativeai